refactor: rename createNodeIpc to createNodeRpc (#3206)
